Don't set Resource.loadergroup in MainFrame.
I think I did this originally to counter-act the effect of the
static{} block that pre-loaded resource being called from a thread
outside the right thread group. Since that static{} block has been
removed and replaced with an appropriate function in MainFrame
and HavenApplet that are always called from the right thread group,
it should not be a problem anymore.

I'm not sure, though. I guess I'll find out if I'm wrong. :)