Try to generally clip color values into their correct range.
I am starting to suspect that the spurious errors for illegal color
values are caused by System.currentTimeMillis moving backwards. I
though it would be guaranteed not to happen, but the Interwebs seem
to corroborate the possiblity.

Clipping the color values should be simpler than clipping time values
every here and there.